@@ -7,55 +7,56 @@
     "start": "ng serve",
     "build": "ng build",
     "test": "ng test",
-    "test:travis-headless": "ng test --single-run --browser=ChromeHeadless -- --no-sandbox",
+    "test:travis-headless": "ng test --watch=false --browsers=ChromeHeadless",
     "lint": "ng lint",
     "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/build-optimizer": "^0.5.7",
-    "@angular/animations": "5.2.9",
-    "@angular/cdk": "5.2.4",
-    "@angular/common": "5.2.9",
-    "@angular/compiler": "5.2.9",
-    "@angular/core": "5.2.9",
-    "@angular/forms": "5.2.9",
-    "@angular/http": "5.2.9",
-    "@angular/material": "5.2.4",
-    "@angular/platform-browser": "5.2.9",
-    "@angular/platform-browser-dynamic": "5.2.9",
-    "@angular/router": "5.2.9",
-    "@ngrx/store": "5.2.0",
-    "@ngrx/store-devtools": "5.2.0",
-    "core-js": "2.5.3",
-    "hammerjs": "2.0.8",
-    "immutable": "3.8.2",
-    "material-design-icons": "3.0.1",
-    "rxjs": "5.5.7",
-    "typed-immutable-record": "0.0.6",
-    "zone.js": "0.8.20"
+    "@angular-devkit/build-optimizer": "^0.7.1",
+    "@angular/animations": "^6.1.0",
+    "@angular/cdk": "^6.4.2",
+    "@angular/common": "^6.1.0",
+    "@angular/compiler": "^6.1.0",
+    "@angular/core": "^6.1.0",
+    "@angular/forms": "^6.1.0",
+    "@angular/http": "^6.1.0",
+    "@angular/material": "^6.4.2",
+    "@angular/platform-browser": "^6.1.0",
+    "@angular/platform-browser-dynamic": "^6.1.0",
+    "@angular/router": "^6.1.0",
+    "@ngrx/store": "^6.0.1",
+    "@ngrx/store-devtools": "^6.0.1",
+    "core-js": "^2.5.4",
+    "hammerjs": "^2.0.8",
+    "immutable": "^3.8.2",
+    "material-design-icons": "^3.0.1",
+    "rxjs": "^6.0.0",
+    "rxjs-tslint": "^0.1.5",
+    "typed-immutable-record": "^0.0.6",
+    "zone.js": "~0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.7.3",
-    "@angular/compiler-cli": "5.2.9",
-    "@angular/language-service": "5.2.9",
-    "@types/jasmine": "2.8.6",
-    "@types/jasminewd2": "2.0.3",
-    "@types/node": "9.4.7",
-    "codelyzer": "4.3.0",
-    "jasmine-core": "2.9.1",
-    "jasmine-spec-reporter": "4.2.1",
-    "karma": "1.7.1",
-    "karma-chrome-launcher": "2.2.0",
-    "karma-cli": "1.0.1",
-    "karma-coverage-istanbul-reporter": "1.4.2",
-    "karma-jasmine": "1.1.1",
-    "karma-jasmine-html-reporter": "0.2.2",
-    "ngrx-store-freeze": "0.2.1",
-    "protractor": "5.3.0",
-    "rxjs-tslint-rules": "^3.17.0",
-    "ts-node": "5.0.1",
-    "tslint": "5.9.1",
-    "typescript": "2.6.2"
+    "@angular-devkit/build-angular": "~0.7.0",
+    "@angular/cli": "^6.1.1",
+    "@angular/compiler-cli": "^6.1.0",
+    "@angular/language-service": "^6.1.0",
+    "@types/jasmine": "~2.8.6",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "~8.9.4",
+    "codelyzer": "~4.2.1",
+    "jasmine-core": "~2.99.1",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~1.7.1",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "~2.0.0",
+    "karma-jasmine": "~1.1.1",
+    "karma-jasmine-html-reporter": "~0.2.2",
+    "ngrx-store-freeze": "^0.2.4",
+    "protractor": "~5.3.0",
+    "rxjs-tslint-rules": "^4.7.1",
+    "ts-node": "~5.0.1",
+    "tslint": "~5.9.1",
+    "typescript": "~2.7.2"
   }
 }
