@@ -5,8 +5,7 @@ import {BoardService} from '../../services/board.service';
 import {Store} from '@ngrx/store';
 import {AppState} from '../../app-store';
 import {blacklistSelector, BoardActions} from '../../model/board/data/board.reducer';
-import {Observable} from 'rxjs/Observable';
-import {Subject} from 'rxjs/Subject';
+import {Observable, Subject} from 'rxjs';
 import {UserSettingActions, userSettingSelector} from '../../model/board/user/user-setting.reducer';
 import {BoardViewModelService} from '../../view-model/board/board-view.service';
 import {BoardHeader} from '../../view-model/board/board-header';
