@@ -14,7 +14,7 @@ import {BoardIssueView} from '../../../../../view-model/board/board-issue-view';
 import {BoardHeader} from '../../../../../view-model/board/board-header';
 import {UpdateParallelTaskEvent} from '../../../../../events/update-parallel-task.event';
 import {IssueDetailState} from '../../../../../model/board/user/issue-detail/issue-detail.model';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 import {ScrollPositionAndHeight} from '../../../../../common/scroll-position-height';
 
 @Component({
