@@ -2,7 +2,8 @@ import {Injectable} from '@angular/core';
 import {Store} from '@ngrx/store';
 import {AppState} from '../../app-store';
 import {BoardViewModelUtil, initialBoardViewModel, initialIssueTable} from './board-view.model';
-import {Observable} from 'rxjs/Observable';
+import {Observable, combineLatest} from 'rxjs';
+import {map} from 'rxjs/operators';
 import {BoardState} from '../../model/board/data/board';
 import {initialUserSettingState} from '../../model/board/user/user-setting.model';
 import {initialBoardState} from '../../model/board/data/board.model';
@@ -41,7 +42,6 @@ import {IssueDetailState} from '../../model/board/user/issue-detail/issue-detail
 import {Dictionary} from '../../common/dictionary';
 import {FontSizeTableService} from '../../services/font-size-table.service';
 import {IssueHeightCalculator} from './issue-height-calculator';
-import {combineLatest} from 'rxjs/observable/combineLatest';
 
 @Injectable()
 export class BoardViewModelService {
@@ -74,61 +74,67 @@ export class BoardViewModelHandler {
 
   getBoardViewModel(boardState$: Observable<BoardState>,
                     userSettingState$: Observable<UserSettingState>): Observable<BoardViewModel> {
-    return combineLatest(boardState$, userSettingState$, (boardState, userSettingState) => {
-      if (boardState === initialBoardState || userSettingState === initialUserSettingState) {
-        return this._lastBoardView;
-      }
 
-      let changeType: ChangeType = null;
-      if (boardState !== this._lastBoardState) {
-        if (this._forcedRefresh) {
-          changeType = ChangeType.UPDATE_BOARD_AFTER_BACKLOG_TOGGLE;
-          this._forcedRefresh = false;
-        } else {
-          if (this._lastBoardState.headers.helpTexts !== boardState.headers.helpTexts) {
-            changeType = ChangeType.INIT_HELP_TEXTS;
-          } else {
-            changeType = this._lastBoardState === initialBoardState ? ChangeType.LOAD_BOARD : ChangeType.UPDATE_BOARD;
+    return combineLatest(boardState$, userSettingState$)
+      .pipe(
+        map((values: any[], index: number) => {
+          const boardState: BoardState = values[0];
+          const userSettingState: UserSettingState = values[1];
+
+          if (boardState === initialBoardState || userSettingState === initialUserSettingState) {
+            return this._lastBoardView;
           }
-        }
-      } else if (boardState.viewId >= 0) {
-        if (userSettingState.filters !== this._lastUserSettingState.filters) {
-          changeType = ChangeType.APPLY_FILTERS;
-        } else if (userSettingState.swimlane !== this._lastUserSettingState.swimlane) {
-          changeType = ChangeType.CHANGE_SWIMLANE;
-        } else if (userSettingState.viewMode !== this._lastUserSettingState.viewMode) {
-          if (this._lastUserSettingState.showBacklog !== userSettingState.showBacklog) {
-            // Don't do anything for this update. The caller needs to do a full refresh (see comment in BoardComponent).
-            this._forcedRefresh = true;
-          } else {
-            changeType = ChangeType.SWITCH_VIEW_MODE;
+
+          let changeType: ChangeType = null;
+          if (boardState !== this._lastBoardState) {
+            if (this._forcedRefresh) {
+              changeType = ChangeType.UPDATE_BOARD_AFTER_BACKLOG_TOGGLE;
+              this._forcedRefresh = false;
+            } else {
+              if (this._lastBoardState.headers.helpTexts !== boardState.headers.helpTexts) {
+                changeType = ChangeType.INIT_HELP_TEXTS;
+              } else {
+                changeType = this._lastBoardState === initialBoardState ? ChangeType.LOAD_BOARD : ChangeType.UPDATE_BOARD;
+              }
+            }
+          } else if (boardState.viewId >= 0) {
+            if (userSettingState.filters !== this._lastUserSettingState.filters) {
+              changeType = ChangeType.APPLY_FILTERS;
+            } else if (userSettingState.swimlane !== this._lastUserSettingState.swimlane) {
+              changeType = ChangeType.CHANGE_SWIMLANE;
+            } else if (userSettingState.viewMode !== this._lastUserSettingState.viewMode) {
+              if (this._lastUserSettingState.showBacklog !== userSettingState.showBacklog) {
+                // Don't do anything for this update. The caller needs to do a full refresh (see comment in BoardComponent).
+                this._forcedRefresh = true;
+              } else {
+                changeType = ChangeType.SWITCH_VIEW_MODE;
+              }
+            } else if (userSettingState.showBacklog !== this._lastUserSettingState.showBacklog) {
+              // Don't do anything. The caller needs to do a full refresh (see comment in BoardComponent).
+              this._forcedRefresh = true;
+            } else if (userSettingState.columnVisibilities !== this._lastUserSettingState.columnVisibilities) {
+              changeType = ChangeType.CHANGE_COLUMN_VISIBILITY;
+            } else if (userSettingState.swimlaneShowEmpty !== this._lastUserSettingState.swimlaneShowEmpty) {
+              changeType = ChangeType.TOGGLE_SWIMLANE_SHOW_EMPTY;
+            } else if (userSettingState.collapsedSwimlanes !== this._lastUserSettingState.collapsedSwimlanes) {
+              changeType = ChangeType.TOGGLE_SWIMLANE_COLLAPSED;
+            } else if (userSettingState.issueDetail !== this._lastUserSettingState.issueDetail) {
+              changeType = ChangeType.UPDATE_ISSUE_DETAIL;
+            }
           }
-        } else if (userSettingState.showBacklog !== this._lastUserSettingState.showBacklog) {
-          // Don't do anything. The caller needs to do a full refresh (see comment in BoardComponent).
-          this._forcedRefresh = true;
-        } else if (userSettingState.columnVisibilities !== this._lastUserSettingState.columnVisibilities) {
-          changeType = ChangeType.CHANGE_COLUMN_VISIBILITY;
-        } else if (userSettingState.swimlaneShowEmpty !== this._lastUserSettingState.swimlaneShowEmpty) {
-          changeType = ChangeType.TOGGLE_SWIMLANE_SHOW_EMPTY;
-        } else if (userSettingState.collapsedSwimlanes !== this._lastUserSettingState.collapsedSwimlanes) {
-          changeType = ChangeType.TOGGLE_SWIMLANE_COLLAPSED;
-        } else if (userSettingState.issueDetail !== this._lastUserSettingState.issueDetail) {
-          changeType = ChangeType.UPDATE_ISSUE_DETAIL;
-        }
-      }
 
-      let boardView: BoardViewModel = this._lastBoardView;
-      if (changeType != null) {
-        boardView = new BoardViewBuilder(this._fontSizeTable, changeType, this._lastBoardView,
-          this._lastBoardState, boardState, this._lastUserSettingState, userSettingState)
-          .build();
-      }
+          let boardView: BoardViewModel = this._lastBoardView;
+          if (changeType != null) {
+            boardView = new BoardViewBuilder(this._fontSizeTable, changeType, this._lastBoardView,
+              this._lastBoardState, boardState, this._lastUserSettingState, userSettingState)
+              .build();
+          }
 
-      this._lastBoardState = boardState;
-      this._lastUserSettingState = userSettingState;
-      this._lastBoardView = boardView;
-      return this._lastBoardView;
-    });
+          this._lastBoardState = boardState;
+          this._lastUserSettingState = userSettingState;
+          this._lastBoardView = boardView;
+          return this._lastBoardView;        })
+      );
   }
 }
 
@@ -1252,7 +1258,7 @@ class RankViewBuilder {
     } else {
       this._currentMap =
         this._existing.reduce(
-          (map, entry) => map.set(entry.issue.key, entry), Map<string, RankViewEntry>());
+          (reduction, entry) => reduction.set(entry.issue.key, entry), Map<string, RankViewEntry>());
     }
   }
 
