@@ -3,8 +3,7 @@ import {FixedHeaderView} from '../fixed-header-view';
 import {BoardHeader} from '../../../../view-model/board/board-header';
 import {BoardViewMode} from '../../../../model/board/user/board-view-mode';
 import {UpdateParallelTaskEvent} from '../../../../events/update-parallel-task.event';
-import {BehaviorSubject} from 'rxjs/BehaviorSubject';
-import {Subject} from 'rxjs/Subject';
+import {BehaviorSubject, Subject} from 'rxjs';
 import {ScrollPositionAndHeightSubject} from '../../../../common/scroll-position-height.subject';
 import {ScrollPositionAndHeight} from '../../../../common/scroll-position-height';
 import {takeUntil} from 'rxjs/operators';
